fixes Code for DoG is duplicated #1509
